+++ src/main/scala/sbtbuildinfo/BuildInfoRenderer.scala
@@ -1,12 +1,14 @@
 package sbtbuildinfo
 
+object BuildInfoRenderer {
+  type Factory = (Seq[BuildInfoOption], String, String) => BuildInfoRenderer
+}
+
 trait BuildInfoRenderer {
 
   def fileType: BuildInfoType
   def extension: String
-  def header: Seq[String]
   def renderKeys(infoKeysNameAndValues: Seq[BuildInfoResult]): Seq[String]
-  def footer: Seq[String]
 
   def isSource = fileType == BuildInfoType.Source
   def isResource = fileType == BuildInfoType.Resource
